requirements.txt
arrow
https://github.com/Rapptz/discord.py/archive/rewrite.zip#egg=discord.py[voice]
eventlet
flask
flask-cors
flask_restful
flask-socketio
gevent
pymongo
pyyaml
requests
